# api.yaml
openapi: 3.0.0
info:
  title: Dice Duel API
  version: 1.1.0
  description: API for multiplayer dice game with round-based mechanics
servers:
  - url: http://localhost:3000/api
paths:
...
  /matches/{matchId}/rounds:
    get:
      summary: Get match rounds
      parameters:
        - $ref: '#/components/parameters/matchId'
      responses:
        '200':
          description: Round list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundList'
  /matches/{matchId}/rounds/{roundId}:
    parameters:
      - $ref: '#/components/parameters/matchId'
      - $ref: '#/components/parameters/roundId'
    get:
      summary: Get round
      responses:
        '200':
          description: Round data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Round'
    put:
      summary: Replace round
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoundRequest'
      responses:
        '200':
          description: Round replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Round'
    patch:
      summary: Patch round
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchRoundRequest'
      responses:
        '200':
          description: Round patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Round'
    delete:
      summary: Delete round
      responses:
        '204':
          description: Round deleted
  /matches/{matchId}/rounds/{roundId}/roll:
    post:
      summary: Roll dice
      parameters:
        - $ref: '#/components/parameters/matchId'
        - $ref: '#/components/parameters/roundId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollDiceRequest'
      responses:
        '200':
          description: Dice rolled
  /matches/{matchId}/rounds/{roundId}/lock:
    post:
      summary: Lock dice
      parameters:
        - $ref: '#/components/parameters/matchId'
        - $ref: '#/components/parameters/roundId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockDiceRequest'
      responses:
        '200':
          description: Dice locked
  /matches/{matchId}/rounds/{roundId}/locked-dice:
    patch:
      summary: Update locked dice
      parameters:
        - $ref: '#/components/parameters/matchId'
        - $ref: '#/components/parameters/roundId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLockedDiceRequest'
      responses:
        '200':
          description: Locked dice updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Round'
  /matches/{matchId}/rounds/{roundId}/resolve:
    post:
      summary: Resolve round
      parameters:
        - $ref: '#/components/parameters/matchId'
        - $ref: '#/components/parameters/roundId'
      responses:
        '200':
          description: Round resolved
...
components:
  parameters:
...
    matchId:
      name: matchId
      in: path
      required: true
      schema:
        type: string
    roundId:
      name: roundId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    RoundStatus:
      type: string
      enum: [INITIALIZED, ROLLING, TARGET_SELECTION, ABILITY_PHASE, RESOLVED]
    DiceFace:
      type: string
      enum: [ATTACK, SHIELD, STEAL]
...
    Round:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/RoundStatus'
        dice:
          type: array
          items:
            $ref: '#/components/schemas/DiceFace'
        locked:
          type: array
          items:
            type: boolean
    RoundList:
      type: array
      items:
        $ref: '#/components/schemas/Round'
    UpdateRoundRequest:
      type: object
      required: [status, dice, locked]
      properties:
        status:
          $ref: '#/components/schemas/RoundStatus'
        dice:
          type: array
          items:
            $ref: '#/components/schemas/DiceFace'
        locked:
          type: array
          items:
            type: boolean
    PatchRoundRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/RoundStatus'
        dice:
          type: array
          items:
            $ref: '#/components/schemas/DiceFace'
        locked:
          type: array
          items:
            type: boolean
    RollDiceRequest:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: string
    LockDiceRequest:
      type: object
      required: [playerId, lockedIndexes]
      properties:
        playerId:
          type: string
        lockedIndexes:
          type: array
          items:
            type: integer
    UpdateLockedDiceRequest:
      type: object
      required: [locked]
      properties:
        locked:
          type: array
          items:
            type: boolean
...